script 경로 수정

+++ scripts/deploy.sh
@@ -1,12 +1,11 @@
 #!/usr/bin/env bash
 
 REPOSITORY=/home/ubuntu/app
-cd $REPOSITORY
-APP_NAME=cidemo
-JAR_NAME=$(ls -t $REPOSITORY/*SNAPSHOT.jar | head -n 1)
 
 echo "> 현재 구동 중인 애플리케이션 pid 확인"
-CURRENT_PID=$(pgrep -fla java | grep $APP_NAME | awk '{print $1}')
+
+CURRENT_PID=$(pgrep -fla java | grep cidemo | awk '{print $1}')
+
 echo "현재 구동 중인 애플리케이션 pid: $CURRENT_PID"
 
 if [ -z "$CURRENT_PID" ]; then
@@ -15,16 +14,28 @@ else
   echo "> kill -9 $CURRENT_PID"
   kill -9 $CURRENT_PID
 
-  while pgrep -fla java | grep $APP_NAME > /dev/null; do
-    echo "기존 프로세스가 종료될 때까지 대기 중..."
-    sleep 1
+  while pgrep -fla java | grep cidemo > /dev/null; do
+      echo "기존 프로세스가 종료될 때까지 대기 중..."
+      sleep 1
   done
   echo "기존 프로세스가 종료되었습니다."
+#  sleep 5
 fi
 
 echo "> 새 애플리케이션 배포"
+
+JAR_NAME=$(ls -tr $REPOSITORY/*SNAPSHOT.jar | tail -n 1)
+
 echo "> JAR NAME: $JAR_NAME"
+if [ ! -f "$JAR_NAME" ]; then
+  echo "실행할 JAR 파일을 찾을 수 없습니다: $JAR_NAME"
+  exit 1
+fi
+
 echo "> $JAR_NAME 에 실행권한 추가"
+
 chmod +x $JAR_NAME
+
 echo "> $JAR_NAME 실행"
+
 nohup java -jar -Duser.timezone=Asia/Seoul $JAR_NAME >> $REPOSITORY/nohup.out 2>&1 &
